src/cardsync/requirements.txt
click==8.1.7
Faker==33.1.0
Flask==3.1.0
flask_bootstrap==3.3.7.1
flask_moment==1.0.6
flask_wtf==1.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
WTForms==3.2.1
